# AgenticInventoryService/AzureAPI/service_swagger_spec_v1.openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "HGV Agentic Inventory Service",
        "description": "API for the HGV Agentic Inventory Service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/jobs/{job_name}": {
            "post": {
                "tags": [
                    "jobs"
                ],
                "summary": "Create Job",
                "description": "Create a new job configuration",
                "operationId": "create_job_api_jobs__job_name__post",
                "parameters": [
                    {
                        "name": "job_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Job Name"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobConfig"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Create Job Api Jobs  Job Name  Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "jobs"
                ],
                "summary": "Describe Job",
                "description": "Get detailed information about a job including its configuration and execution history",
                "operationId": "describe_job_api_jobs__job_name__get",
                "parameters": [
                    {
                        "name": "job_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Job Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "jobs"
                ],
                "summary": "Delete Job",
                "description": "Delete a job configuration",
                "operationId": "delete_job_api_jobs__job_name__delete",
                "parameters": [
                    {
                        "name": "job_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Job Name"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/jobs/{job_name}/status": {
            "get": {
                "tags": [
                    "jobs"
                ],
                "summary": "Get Job Status",
                "description": "Get the status of a specific job",
                "operationId": "get_job_status_api_jobs__job_name__status_get",
                "parameters": [
                    {
                        "name": "job_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Job Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Job Status Api Jobs  Job Name  Status Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/jobs/{job_name}/trigger": {
            "post": {
                "tags": [
                    "jobs"
                ],
                "summary": "Trigger Job",
                "description": "Trigger a job execution",
                "operationId": "trigger_job_api_jobs__job_name__trigger_post",
                "parameters": [
                    {
                        "name": "job_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Job Name"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Trigger Job Api Jobs  Job Name  Trigger Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/jobs/{job_name}/cancel": {
            "post": {
                "tags": [
                    "jobs"
                ],
                "summary": "Cancel Job",
                "description": "Cancel a running job",
                "operationId": "cancel_job_api_jobs__job_name__cancel_post",
                "parameters": [
                    {
                        "name": "job_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Job Name"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Cancel Job Api Jobs  Job Name  Cancel Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "JobConfig": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "job_name": {
                        "type": "string",
                        "title": "Job Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "schedule": {
                        "type": "string",
                        "title": "Schedule"
                    },
                    "parameters": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Parameters"
                    },
                    "output_location": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Output Location"
                    },
                    "notify": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/NotifyConfig"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "version": {
                        "type": "string",
                        "title": "Version",
                        "default": "1.0.0"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags"
                    },
                    "enabled": {
                        "type": "boolean",
                        "title": "Enabled",
                        "default": true
                    },
                    "timeout": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Timeout"
                    },
                    "retry_count": {
                        "type": "integer",
                        "title": "Retry Count",
                        "default": 3
                    },
                    "retry_delay": {
                        "type": "integer",
                        "title": "Retry Delay",
                        "default": 60
                    },
                    "jitter": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Jitter",
                        "description": "Random jitter in seconds to prevent  jobs from starting at exactly the same time."
                    }
                },
                "type": "object",
                "required": [
                    "job_name",
                    "schedule"
                ],
                "title": "JobConfig",
                "description": "Model for job configuration stored in YAML."
            },
            "NotifyConfig": {
                "properties": {
                    "enabled": {
                        "type": "boolean",
                        "title": "Enabled",
                        "default": true
                    },
                    "method": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Method"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "webhook": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Webhook"
                    }
                },
                "type": "object",
                "title": "NotifyConfig",
                "description": "Model for job notification configuration."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}